requirements.txt
boto3==1.26.59
django-storages==1.13.2
gunicorn==20.1.0
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
idna==3.3
Pillow==9.1.0
psycopg2-binary
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata
urllib3
whitenoise==6.1.0
django-environ==0.9.0
django-cors-headers==4.0.0
mistralai==0.1.8
djangorestframework_simplejwt